requirements.txt
torch
torchvision
torchaudio
opensmile
ffmpeg-python
pyloudnorm
